// src/components/forms/MultiStepSignUpForm/Step1/SignUpFormInput.Schema.tsx
import * as yup from "yup";

export const signUpSchema = yup.object().shape({
  avatarId: yup.number().required("Avatar is required"),
  firstName: yup.string().required("First name is required"),
  lastName: yup.string().required("Last name is required"),
  userName: yup.string().required("Username is required"),
  email: yup.string().email("Invalid email").required("Email is required"),
  password: yup
    .string()
    .required("Password is required")
    .min(8, "Password must be at least 8 characters"),
});